+++ Server/Server/WebSocket/Server.cs
@@ -233,9 +233,9 @@ namespace WebSocket
         /// </summary>
         private void handleGameEnd(Game game)
         {
-            (int,int) scores = game.GetScore();
-            int scorePlayer1 = scores.Item1;
-            int scorePlayer2 = scores.Item2;
+            (float , float) scores = game.GetScore();
+            float scorePlayer1 = scores.Item1;
+            float scorePlayer2 = scores.Item2;
             bool player1won = false;
             bool player2won = false;
